// src/auth/provider/schema.ts
import * as z from "zod";

export const appleProfileSchema = z.object({
  iss: z.literal("https://appleid.apple.com"),
  aud: z.string(),
  iat: z.number(),
  exp: z.number(),
  sub: z.string(),
  nonce: z.string(),
  nonce_supported: z.boolean(),
  email: z.string(),
  email_verified: z.union([z.literal("true"), z.literal(true)]),
  is_private_email: z.union([
    z.literal("true"),
    z.literal("false"),
    z.boolean(),
  ]),
  real_user_status: z.union([z.literal(0), z.literal(1), z.literal(2)]),
  transfer_sub: z.string(),
  at_hash: z.string(),
  auth_time: z.number(),
});

export const githubProfileSchema = z.object({
  login: z.string(),
  id: z.number(),
  node_id: z.string(),
  avatar_url: z.string(),
  gravatar_id: z.string().nullable(),
  url: z.string(),
  html_url: z.string(),
  followers_url: z.string(),
  following_url: z.string(),
  gists_url: z.string(),
  starred_url: z.string(),
  subscriptions_url: z.string(),
  organizations_url: z.string(),
  repos_url: z.string(),
  events_url: z.string(),
  received_events_url: z.string(),
  type: z.string(),
  site_admin: z.boolean(),
  name: z.string().nullable(),
  company: z.string().nullable(),
  blog: z.string().nullable(),
  location: z.string().nullable(),
  email: z.string().nullable(),
  hireable: z.boolean().nullable(),
  bio: z.string().nullable(),
  twitter_username: z.string().nullable(),
  public_repos: z.number(),
  public_gists: z.number(),
  followers: z.number(),
  following: z.number(),
  created_at: z.string(),
  updated_at: z.string(),
  private_gists: z.number(),
  total_private_repos: z.number(),
  owned_private_repos: z.number(),
  disk_usage: z.number(),
  suspended_at: z.string().nullable().optional(),
  collaborators: z.number(),
  two_factor_authentication: z.boolean(),
  plan: z.object({
    collaborators: z.number(),
    name: z.string(),
    space: z.number(),
    private_repos: z.number(),
  }),
});

export const googleProfileSchema = z.object({
  aud: z.string(),
  azp: z.string(),
  email: z.string(),
  email_verified: z.boolean(),
  exp: z.number(),
  family_name: z.string(),
  given_name: z.string(),
  hd: z.string(),
  iat: z.number(),
  iss: z.string(),
  jti: z.string(),
  name: z.string(),
  nbf: z.number(),
  picture: z.string(),
  sub: z.string(),
});